requirements.txt
selenium
webdriver-manager>=4.0.1
pandas
requests
beautifulsoup4
snowflake-connector-python
spacy
scikit-learn